Minor viz refactor
